Move `GasOperation` into precharge.rs
